Add 'timeOfCollision' member to Ball class
In certain circumstances multiple collisions were detected when only
one was desired to have been detected. This happened when a Ball
collided with a Paddle, and then the new direction of the Ball was
sufficiently focused in the y direction along with the Paddle also
moving in the same y direction such that the Paddle was 'pushing'
the Ball along.

Such motion would be more realistic but isn't desired for this case,
so an extra condition on Ball-Paddle collision detection was added,
stating that there be at least a 1 second gap since the last
Ball-Paddle collision.