Edit Ball's 'calculateVelocityAfterPaddleCollision' function
The current way for calculating the angle a Ball should bounce off
a Paddle after colliding with it is based on finding the y axis
value of the point where the Ball collided with the Paddle.

Away from the top and bottom of the Paddle this works
as intended, but when collisions occur at the top/bottom
of the Paddle the angle of the trajectory after collision
can be very sharp.

As an example, the Ball may hit the top of the Paddle rather
than the left/right, but since the Ball's position is measured
from the top of the Ball rather than the centre this causes
the difference between the Ball's y coord and the Paddle's
y coord to be negative. Due to the calculation of the 'bounce_angle'
variable this then causes the Ball's angle to be much smaller
than the lower limit that is set (this lower limit is set by
assuming the smallest possible value for 'collision_y_coord' is 0,
and so the smallest value 'bounce_angle' can be in degress should
be 90 - 4 * (player.h/2 - 0))

This problem is alleviated by checking if 'collision_y_coord'
exceeds the lower/upper limit and, if so, simply sets it to be
the value of the corresponding boundary, thus assuring that
'bounce_angle' won't be lower/higher than the limits set.