Add constant to Ball's 'calculateVelocityAfterPaddleCollision' function
A value of 4 was used as the step-number of degrees in
Ball's 'calculateVelocityAfterPaddleCollision' function, and
was chosen simply because using that value (in conjunction
with 'player.h' being 30) caused the minimum angle of a Ball
after colliding with a Paddle to be 30 degrees, and the maximum
to be 150 degrees (relative to a vertical line) which seemed
like a reasonable starting point.

However it wasn't very clear why a value of 4 seemingly randomly
appeared in the formula for 'bounce_angle', so hopefully this reduces
any confusion regarding that formula.