Add extra citations to files
LazyFoo's tutorials on SDL was where I learned the limited knowledge
I have about game and graphics programming, and so there are certainly
parts of all the files in this project that started out as sections of
code from those tutorials.

During the project I sometimes forgot to add citations, so this commit
is correcting that.